pyproject.toml
[tool.poetry]
name = "nonebot_plugin_maze"
version = "0.2.2"
description = "NoneBot2交互式解迷宫插件"
authors = ["T_EtherLeaF <thetapilla@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = ">=3.8"
Pillow = "^9.0.0"
asyncer = "^0.0.1"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
